# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Warehouse Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user from session token
      parameters:
        - name: authorization
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List all materials
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create a new material (master only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      summary: Get material by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "404":
          description: Not found
    patch:
      operationId: updateMaterial
      tags: [materials]
      summary: Update a material (master only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material (master only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /history:
    get:
      operationId: listHistory
      tags: [history]
      summary: List all scan history (in and out)
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [in, out]
        - name: materialId
          in: query
          required: false
          schema:
            type: integer
        - name: userId
          in: query
          required: false
          schema:
            type: integer
        - name: from
          in: query
          required: false
          schema:
            type: string
        - name: to
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: History records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryRecord"
  /history/{id}:
    delete:
      operationId: deleteHistory
      tags: [history]
      summary: Delete a history record (master only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [master, user]
        createdAt:
          type: string
...
    Material:
      type: object
      required: [id, name, code, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    MaterialInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
    MaterialUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
...
    HistoryRecord:
      type: object
      required: [id, type, materialId, materialName, boxLabel, userId, userName, serialNumbers, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [in, out]
        materialId:
          type: ["integer", "null"]
        materialName:
          type: ["string", "null"]
        boxLabel:
          type: ["string", "null"]
        userId:
          type: integer
        userName:
          type: string
        serialNumbers:
          type: array
          items:
            type: string
        createdAt:
          type: string
...