# lib/api-zod/src/generated/types/materialInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Warehouse Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface MaterialInput {
  name: string;
  code: string;
  description?: string;
}